🔨 cdn处理逻辑代码重构
