+++ package/vite-add-cdn-script/lib/utils/cdnProcess/index.ts
@@ -1,0 +1,1 @@
+export * from "./lib";
